Refaktorering av skjemautlistingen
