+++ src/sider/interne/skjemautlisting/Oversikt.tsx
@@ -3,13 +3,13 @@ import {
   apiKallAlleSEDSkjemaer,
   apiKallAlleSkjemaer
 } from "../../../klienter/sanityKlient";
-import Skjematabell from "./Tabell";
 import Hovedbanner from "../../../komponenter/bannere/Hovedbanner";
 import { Skjema } from "../../../typer/skjemaogvedlegg";
 import PanelBase from "nav-frontend-paneler";
 import { RouteComponentProps } from "react-router";
 import { injectIntl, InjectedIntlProps, FormattedMessage } from "react-intl";
 import Lenkepanel from "nav-frontend-lenkepanel/lib";
+import Oversiktstabell from "./tabeller/Oversiktstabell";
 
 interface State {
   skjemaer: Skjema[];
@@ -23,7 +23,6 @@ class SkjemautlistingOversikt extends React.Component<Props, State> {
   }
 
   componentDidUpdate(prevProps: RouteComponentProps<{ skjematype?: string }>) {
-    // Typical usage (don't forget to compare props):
     if (
       this.props.match.params.skjematype !== prevProps.match.params.skjematype
     ) {
@@ -41,30 +40,6 @@ class SkjemautlistingOversikt extends React.Component<Props, State> {
         this.setState({ skjemaer: sanityData });
       });
     }
-  };
-
-  grupperEmneordOgSkjemaer = (skjemaer: Skjema[]) => {
-    const emneordMedSkjemaer: { [emneord: string]: Skjema[] } = {};
-    const skjemaUtenEmneord = "Skjemaer uten emneord";
-    const skjemaliste: { skjemanummer: string; skjemanavn: string }[] = [];
-
-    skjemaer.map(skjema => {
-      skjemaliste.push({
-        skjemanummer: skjema.skjemanummer,
-        skjemanavn: skjema.navn ? skjema.navn.nb || "" : ""
-      });
-      return skjema.emneord
-        ? skjema.emneord.map(emneord =>
-            !emneordMedSkjemaer.hasOwnProperty(emneord.emneord)
-              ? (emneordMedSkjemaer[emneord.emneord] = [skjema])
-              : emneordMedSkjemaer[emneord.emneord].push(skjema)
-          )
-        : !emneordMedSkjemaer.hasOwnProperty(skjemaUtenEmneord)
-        ? (emneordMedSkjemaer[skjemaUtenEmneord] = [skjema])
-        : emneordMedSkjemaer[skjemaUtenEmneord].push(skjema);
-    });
-
-    return <Skjematabell data={emneordMedSkjemaer} skjemaliste={skjemaliste} />;
   };
 
   render = () => {
@@ -97,7 +72,7 @@ class SkjemautlistingOversikt extends React.Component<Props, State> {
               </Lenkepanel>
             </div>
             {this.state
-              ? this.grupperEmneordOgSkjemaer(this.state.skjemaer)
+              ? <Oversiktstabell skjemaer={this.state.skjemaer}/>
               : null}
           </PanelBase>
         </div>
